move app component
