+++ frontend/src/components/App.tsx
@@ -1,11 +1,11 @@
 import {ReactQueryDevtools} from 'react-query/devtools'
 import React, {useContext, useEffect} from "react";
-import ExperimentTrial from "../Trials";
-import {NetworkContextProvider} from "../../contexts/NetworkContext";
+import ExperimentTrial from "./Trials";
+import {NetworkContextProvider} from "../contexts/NetworkContext";
 import {QueryClient, QueryClientProvider} from 'react-query';
 import {useSearchParams} from "react-router-dom";
 import {v4 as uuid4} from "uuid";
-import SessionContextProvider from "../../contexts/SessionContext";
+import SessionContextProvider from "../contexts/SessionContext";
 
 
 // Create a client
